+++ installation-bundle/src/Database/Version450Update.php
@@ -43,6 +43,20 @@ class Version450Update extends AbstractVersionUpdate
         ');
 
         $this->connection->query("
+            ALTER TABLE
+                tl_content
+            ADD
+                youtubeStart int(10) unsigned NOT NULL default '0'
+        ");
+
+        $this->connection->query("
+            ALTER TABLE
+                tl_content
+            ADD
+                youtubeStop int(10) unsigned NOT NULL default '0'
+        ");
+
+        $this->connection->query("
             UPDATE
                 tl_form_field
             SET