+++ news-bundle/src/Resources/contao/modules/ModuleNewsList.php
@@ -58,6 +58,12 @@ class ModuleNewsList extends ModuleNews
 			return '';
 		}
 
+		// Show the news reader if an item has been selected
+		if ($this->news_readerModule > 0 && (isset($_GET['items']) || (\Config::get('useAutoItem') && isset($_GET['auto_item']))))
+		{
+			return $this->getFrontendModule($this->news_readerModule, $this->strColumn);
+		}
+
 		return parent::generate();
 	}